requirements.txt
# Core Dependencies - VERSÕES COMPATÍVEIS
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.5
langchain-core>=0.3.17,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0
openai==1.54.4

# Fix compatibilidade: httpx 0.28 removeu 'proxies'; manter <0.28
httpx<0.28,>=0.23.0

# Web Framework
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.6.0

# HTTP & API
requests==2.31.0

# Database & Storage
redis==5.0.1
psycopg==3.2.12

# AI & ML
cohere==4.47

# Utilities
python-dotenv==1.0.0
pytz==2024.1

# Optional: Logging & Monitoring
python-json-logger==2.0.7
